Move Statistics model to Services and update references
