+++ src/Controller/Admin/Ajax/Stats/FinanceStats.php
@@ -4,14 +4,14 @@
  * HugaShop - Selling anything
  *
  * @author Andri Huga
- * @version 2.3
+ * @version 2.4
  *
  */
 
 namespace App\Controller\Admin\Ajax\Stats;
 
 use HugaShop\Services\Request;
-use HugaShop\Models\Statistics;
+use HugaShop\Services\Statistics;
 use App\Controller\BaseAdminController;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
