+++ src/Controller/Admin/Ajax/Stats/OrderStats.php
@@ -4,14 +4,14 @@
  * HugaShop - Selling anything
  *
  * @author Andri Huga
- * @version 2.2
+ * @version 2.3
  *
  */
 
 namespace App\Controller\Admin\Ajax\Stats;
 
 use HugaShop\Services\Request;
-use HugaShop\Models\Statistics;
+use HugaShop\Services\Statistics;
 use HugaShop\Models\Finance\FinancePayment;
 use App\Controller\BaseAdminController;
 use Symfony\Component\Routing\Attribute\Route;
